minor fixes | refactor governance
